requirements.txt
fastapi
uvicorn
opencv-python
numpy
face_recognition
requests
pillow
